refactor(shared): mueve los guards de auth de actions/ a shared/lib/serverAuth.ts
