+++ src/actions/archive/saveArchivedNotes.ts
@@ -2,7 +2,7 @@
 
 import { prisma } from '@/shared/lib/prisma'
 import type { LibraryOfArchivedNotes } from '@/modules/notes/LibraryOfArchiveNotes/model/libraryOfArchivedNotes'
-import { requireBoardAccess } from '../shared'
+import { requireBoardAccess } from '@/shared/lib/serverAuth'
 
 export async function saveArchivedNotes({
 	boardId,
