+++ src/actions/board/deleteBoard.ts
@@ -1,7 +1,7 @@
 'use server'
 
 import { prisma } from '@/shared/lib/prisma'
-import { requireBoardAccess } from '../shared'
+import { requireBoardAccess } from '@/shared/lib/serverAuth'
 
 export async function deleteBoard({ boardId }: { boardId: string }): Promise<void> {
 	await requireBoardAccess(boardId)
