+++ src/actions/board/getBoards.ts
@@ -2,7 +2,7 @@
 
 import { prisma } from '@/shared/lib/prisma'
 import type { Board } from '@/modules/Dashboard/model/board'
-import { requireAuth } from '../auth'
+import { requireAuth } from '@/shared/lib/serverAuth'
 
 export async function getBoards(): Promise<Board[]> {
 	const userId = await requireAuth()
